+++ CHANGELOG.md
@@ -4,7 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
-### 🔄 Changed
+### ✅ Added
+- SDK version info sent in all requests
+- Call participants shown in the lobby view
 
 # [0.1.0](https://github.com/GetStream/stream-video-swift/releases/tag/0.1.0)
 _July 07, 2023_